Clean up irrlicht includes; avoiding including the huge irrlicht.h when we only need one specific class, especially in header files; use forwards when possible
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@8660 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/graphics/material_manager.cpp
@@ -227,7 +227,7 @@ Material *MaterialManager::getMaterial(const std::string& fname,
     m_materials.push_back(m);
     if(make_permanent) 
     {
-        assert(m_shared_material_index==m_materials.size()-1);
+        assert(m_shared_material_index==(int)m_materials.size()-1);
         m_shared_material_index = (int)m_materials.size();
     }
     return m ;
